test: move all common code to DhisControllerTestBase
