+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/DhisControllerConvenienceTest.java
@@ -27,31 +27,8 @@
  */
 package org.hisp.dhis.webapi;
 
-import java.util.Date;
-import lombok.Getter;
 import org.hisp.dhis.IntegrationH2Test;
-import org.hisp.dhis.common.CodeGenerator;
-import org.hisp.dhis.common.IdentifiableObjectManager;
-import org.hisp.dhis.config.H2DhisConfiguration;
-import org.hisp.dhis.dbms.DbmsManager;
-import org.hisp.dhis.render.RenderService;
-import org.hisp.dhis.user.User;
-import org.hisp.dhis.user.UserDetails;
-import org.hisp.dhis.user.UserRole;
-import org.hisp.dhis.user.UserService;
-import org.hisp.dhis.utils.TestUtils;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.transaction.TransactionDefinition;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionTemplate;
-import org.springframework.web.context.WebApplicationContext;
 
 /**
  * Base class for convenient testing of the web API on basis of {@link
@@ -59,87 +36,6 @@ import org.springframework.web.context.WebApplicationContext;
  *
  * @author Jan Bernitt
  */
-@ExtendWith(SpringExtension.class)
-@WebAppConfiguration
-@ContextConfiguration(
-    classes = {H2DhisConfiguration.class, MvcTestConfig.class, WebTestConfiguration.class})
 @ActiveProfiles("test-h2")
 @IntegrationH2Test
-@Transactional
-public abstract class DhisControllerConvenienceTest extends DhisControllerTestBase {
-  @Autowired private WebApplicationContext webApplicationContext;
-
-  @Autowired private UserService _userService;
-
-  @Autowired private RenderService _renderService;
-
-  @Autowired protected IdentifiableObjectManager manager;
-
-  @Autowired protected DbmsManager dbmsManager;
-
-  @Autowired private TransactionTemplate txTemplate;
-
-  @Getter private User adminUser;
-
-  @BeforeEach
-  final void setup() {
-    userService = _userService;
-    renderService = _renderService;
-    clearSecurityContext();
-
-    this.adminUser = _preCreateInjectAdminUserWithoutPersistence();
-    manager.persist(adminUser);
-    _injectSecurityContextUser(adminUser);
-
-    superUser = createAndAddAdminUser("ALL");
-
-    mvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).build();
-
-    switchContextToUser(superUser);
-    currentUser = superUser;
-
-    TestUtils.executeStartupRoutines(webApplicationContext);
-
-    dbmsManager.flushSession();
-    dbmsManager.clearSession();
-  }
-
-  private void _injectSecurityContextUser(User user) {
-    if (user == null) {
-      clearSecurityContext();
-      return;
-    }
-    hibernateService.flushSession();
-    User user1 = manager.find(User.class, user.getId());
-    injectSecurityContext(UserDetails.fromUser(user1));
-  }
-
-  private User _preCreateInjectAdminUserWithoutPersistence() {
-    UserRole role = createUserRole("Superuser_Test_" + CodeGenerator.generateUid(), "ALL");
-    role.setUid(CodeGenerator.generateUid());
-    manager.persist(role);
-    User user = new User();
-    String uid = CodeGenerator.generateUid();
-    user.setUid(uid);
-    user.setFirstName("Firstname_" + uid);
-    user.setSurname("Surname_" + uid);
-    user.setUsername(DEFAULT_USERNAME + "_test_" + CodeGenerator.generateUid());
-    user.setPassword(DEFAULT_ADMIN_PASSWORD);
-    user.getUserRoles().add(role);
-    user.setLastUpdated(new Date());
-    user.setCreated(new Date());
-    return user;
-  }
-
-  protected final void doInTransaction(Runnable operation) {
-    final int defaultPropagationBehaviour = txTemplate.getPropagationBehavior();
-    txTemplate.setPropagationBehavior(TransactionDefinition.PROPAGATION_REQUIRES_NEW);
-    txTemplate.execute(
-        status -> {
-          operation.run();
-          return null;
-        });
-    // restore original propagation behaviour
-    txTemplate.setPropagationBehavior(defaultPropagationBehaviour);
-  }
-}
+public abstract class DhisControllerConvenienceTest extends DhisControllerTestBase {}
