+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/AbstractCrudControllerTest.java
@@ -88,7 +88,7 @@ class AbstractCrudControllerTest extends DhisControllerConvenienceTest {
         GET("/users/").content(HttpStatus.OK).getList("users", JsonUser.class);
     assertEquals(2, users.size());
     JsonUser user = users.get(0);
-    assertStartsWith("Admin", user.getDisplayName());
+    assertStartsWith("First", user.getDisplayName());
   }
 
   @Test
