+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/UserControllerTest.java
@@ -668,7 +668,7 @@ class UserControllerTest extends DhisControllerConvenienceTest {
 
     manager.flush();
     manager.clear();
-    injectSecurityContextUser(this.superUser);
+    injectSecurityContextUser(getAdminUser());
 
     // assert lastUpdated has been updated by new user & users not empty
     JsonUserGroup userGroupUserAdded =
